requirements.txt
discord.py
image
fabric~=2.7.1
texttable~=1.6.7
aiohttp~=3.8.3
Pillow~=9.3.0